GetNumberOfImages() from int to long return value
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10848 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ MMDevice/DeviceBase.h
@@ -1549,7 +1549,7 @@ protected:
       MM::MMTime GetActualDuration(){return actualDuration_;}
 
       CCameraBase* GetCamera() {return camera_;}
-      int GetNumberOfImages() {return numImages_;}
+      long GetNumberOfImages() {return numImages_;}
 
       void UpdateActualDuration() {actualDuration_ = camera_->GetCurrentMMTime() - startTime_;}
